\name{numerov-package}
\alias{numerov-package}
\alias{numerov}
\docType{package}
\title{
\packageTitle{numerov}
}
\description{
\packageDescription{numerov}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{numerov}
\packageIndices{numerov}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{numerov}

Maintainer: \packageMaintainer{numerov}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
